SeamFind Activation for Fish-eye Cameras on Equator
PTgui calibration modulates the Pitch parameters for all the images for
every new calibration, which breaks our assumption that the pitch
tolerance for a equator camera should be less than 5 degrees, as PTgui
calibration produces pitch which ranges 0 - 30 degree for cameras placed
on equator (Example - Z cam S1). This turns off the seam-find.

Current
logic - Looks at the individual pitch values and checks against the
pitch tolerance; if the pitch of all the cameras on the equator is
lesser than the tolerance, then the special case is turned on.

New
Logic proposed - looks at the difference between the max and min pitch
of the all the cameras and compares it to the pitch tolerance, now we
can compensate for pitch modulated by PTgui.